+++ packages/react/src/form/error-message/index.tsx
@@ -1,0 +1,8 @@
+import "@postenbring/hedwig-css/dist/fonts.css";
+import "@postenbring/hedwig-css/dist/tokens.css";
+import "@postenbring/hedwig-css/dist/body.css";
+
+import "@postenbring/hedwig-css/dist/error-message.css";
+
+export { ErrorMessage } from "./error-message";
+export type * from "./error-message";